fix(eveyrthing): upgrade to NativeScript 3.0.0
This release is almost a full rewrite and fixes a number of issues as well as being made compatible with NativeScript 3.

Close #8, close #15, close #16, close #18. Also related is #17.

BREAKING CHANGE: The plugin no longer auto-registers a tag in Angular applications. Instructions to register a `PDFView` element have been added to the project README.